+++ internal/impl/mongodb/output.go
@@ -75,7 +75,7 @@ func init() {
 					"upsert",
 					"The upsert setting is optional and only applies for update-one and replace-one operations. If the filter specified in filter_map matches,"+
 						"the document is updated or replaced accordingly, otherwise it is created.",
-				).HasDefault(false).HasType(docs.FieldTypeBool),
+				).HasDefault(false).HasType(docs.FieldTypeBool).AtVersion("3.60.0"),
 				docs.FieldCommon(
 					"max_in_flight",
 					"The maximum number of messages to have in flight at a given time. Increase this to improve throughput."),
@@ -172,7 +172,7 @@ func NewWriter(
 		return nil, fmt.Errorf("mongodb hint_map not allowed for '%s' operation", conf.Operation)
 	}
 
-	if upsertAllowed && conf.Upsert != false {
+	if upsertAllowed && conf.Upsert {
 		return nil, fmt.Errorf("mongodb upsert not allowed for '%s' operation", conf.Operation)
 	}